+++ app/src/main/java/com/survivalcoding/todolist/domain/OrderBy.kt
@@ -1,0 +1,8 @@
+package com.survivalcoding.todolist.domain
+
+enum class OrderBy {
+    TIME_ASC,
+    TIME_DESC,
+    TITLE_ASC,
+    TITLE_DESC
+}
